+++ scripts/build.js
@@ -1,28 +1,24 @@
 const fs = require('fs')
-const brotliSize = require('brotli-size');
+const brotliSize = require('brotli-size')
 
-([
-    'main',
-]).forEach(packageName => {
-    if (!fs.existsSync(`./packages/${packageName}/dist`)) {
-        fs.mkdirSync(`./packages/${packageName}/dist`, '0744')
-    }
+if (!fs.existsSync('./dist')) {
+    fs.mkdirSync('./dist', '0744')
+}
 
-    // Go through each file in the packageName's "build" directory
-    // and use the appropriate bundling strategy based on its name.
-    fs.readdirSync(`./packages/${packageName}/builds`).forEach(file => {
-        bundleFile(packageName, file)
-    })
+// Go through each file in the src's "build" directory
+// and use the appropriate bundling strategy based on its name.
+fs.readdirSync('./builds').forEach(file => {
+    bundleFile(file)
 })
 
-function bundleFile(packageName, file) {
+function bundleFile(file) {
     // Based on the filename, give esbuild a specific configuration to build.
     ({
         // This output file is meant to be loaded in a browser's <script> tag.
         'cdn.js': () => {
             build({
-                entryPoints: [`packages/${packageName}/builds/${file}`],
-                outfile: `packages/${packageName}/dist/${file}`,
+                entryPoints: [`builds/${file}`],
+                outfile: `dist/${file}`,
                 bundle: true,
                 platform: 'browser',
                 define: { CDN: true },
@@ -30,14 +26,15 @@ function bundleFile(packageName, file) {
 
             // Build a minified version.
             build({
-                entryPoints: [`packages/${packageName}/builds/${file}`],
-                outfile: `packages/${packageName}/dist/${file.replace('.js', '.min.js')}`,
+                entryPoints: [`builds/${file}`],
+                outfile: `dist/${file.replace('.js', '.min.js')}`,
                 bundle: true,
                 minify: true,
                 platform: 'browser',
                 define: { CDN: true },
             }).then(() => {
-                outputSize(packageName, `packages/${packageName}/dist/${file.replace('.js', '.min.js')}`)
+                outputSize(
+                    `dist/${file.replace('.js', '.min.js')}`)
             })
         },
         // This file outputs two files: an esm module and a cjs module.
@@ -45,16 +42,29 @@ function bundleFile(packageName, file) {
         // and the cjs one is meant for "require" statements (node).
         'module.js': () => {
             build({
-                entryPoints: [`packages/${packageName}/builds/${file}`],
-                outfile: `packages/${packageName}/dist/${file.replace('.js', '.esm.js')}`,
+                entryPoints: [`builds/${file}`],
+                outfile: `dist/${file.replace('.js', '.esm.js')}`,
                 bundle: true,
                 platform: 'neutral',
                 mainFields: ['main', 'module'],
             })
 
+            // Minified ESM file
             build({
-                entryPoints: [`packages/${packageName}/builds/${file}`],
-                outfile: `packages/${packageName}/dist/${file.replace('.js', '.cjs.js')}`,
+                entryPoints: [`builds/${file}`],
+                outfile: `dist/${file.replace('.js', '.esm.min.js')}`,
+                bundle: true,
+                minify: true,
+                platform: 'neutral',
+                mainFields: ['main', 'module'],
+            }).then(() => {
+                outputSize(
+                    `dist/${file.replace('.js', '.esm.min.js')}`)
+            })
+
+            build({
+                entryPoints: [`builds/${file}`],
+                outfile: `dist/${file.replace('.js', '.cjs.js')}`,
                 bundle: true,
                 target: ['node10.4'],
                 platform: 'node',
@@ -73,10 +83,10 @@ function build(options) {
     }).catch(() => process.exit(1))
 }
 
-function outputSize(packageName, file) {
+function outputSize(file) {
     const size = bytesToSize(brotliSize.sync(fs.readFileSync(file)))
 
-    console.log('\x1b[32m', `${packageName}: ${size}`)
+    console.log('\x1b[32m', file + ' : ' + size)
 }
 
 function bytesToSize(bytes) {
@@ -84,5 +94,5 @@ function bytesToSize(bytes) {
     if (bytes === 0) return 'n/a'
     const i = parseInt(Math.floor(Math.log(bytes) / Math.log(1024)), 10)
     if (i === 0) return `${bytes} ${sizes[i]}`
-    return `${(bytes / (1024 ** i)).toFixed(1)} ${sizes[i]}`
+    return `${(bytes / (1024 ** i)).toFixed(1)} ${sizes[i]} `
 }