+++ CHANGELOG.md
@@ -8,6 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+## [2.4.2] - 2024-02-22
+
+### Changed
+
+-   Change build script and include README in NPM package.
+
+## Added
+
+-   Minified ESM module file `dist/module.esm.min.js`.
+
 ## [2.4.1] - 2023-12-28
 
 ### Fixed
@@ -113,3 +123,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 [2.3.0]: https://github.com/pinecone-router/router/compare/2.2.1...2.3.0
 [2.4.0]: https://github.com/pinecone-router/router/compare/2.3.0...2.4.0
 [2.4.1]: https://github.com/pinecone-router/router/compare/2.4.0...2.4.1
+[2.4.2]: https://github.com/pinecone-router/router/compare/2.4.1...2.4.2